Migrate QueryBuilderJson to TS.
